+++ game/gui/components/button.py
@@ -10,9 +10,11 @@ class Button(GUIComponent):
         super().__init__(center_x, center_y)
 
         self.icon = arcade.Sprite(BUTTON_SPRITE, SCALING)
-        self.text = ''
+        self.text = None
+        self.text_settings = {'x': 0, 'y': 0, 'color': arcade.csscolor.WHITE_SMOKE, 'font_size': 0}
         self.center_x = center_x
         self.center_y = center_y
+        self.pressed = False
 
     def set_text(self, text):
         self.text = str(text)
@@ -25,11 +27,14 @@ class Button(GUIComponent):
         if self.pressed:
             self.on_release()
 
+    def on_press(self):
+        pass
+
+    def on_release(self):
+        if self.pressed:
+            self.pressed = False
+
     def check_mouse_press(self, x, y):
-
-        #print("lol checking")
-        #print(self.center_x, self.center_y, self.icon.width, self.icon.height, x, y)
-
         if x > self.center_x + self.icon.width / 2:
             return False
         if x < self.center_x - self.icon.width / 2:
@@ -40,7 +45,11 @@ class Button(GUIComponent):
             return False
         return True
 
-
-
     def draw(self):
-        pass
+        self.icon.center_x = self.center_x
+        self.icon.center_y = self.center_y
+        self.icon.draw()
+
+        arcade.draw_text(self.text,
+                         self.icon.center_x - self.text_settings['x'], self.icon.center_y - self.text_settings['y'],
+                         self.text_settings['color'], self.text_settings['font_size'])
