+++ game/gui/components/coins_label.py
@@ -6,6 +6,5 @@ class CoinsLabel(ScoreLabel):
 	def __init__(self, icon_type, center_x, center_y):
 		super().__init__(icon_type, center_x, center_y)
 
-
 	def update(self):
 		self.set_text(ScoreManager.score)
