+++ game/gui/menu_components/menu_panel.py
@@ -1,0 +1,11 @@
+from game.gui.panel import GUIPanel
+
+
+class MenuPanel(GUIPanel):
+    def __init__(self):
+        super().__init__()
+        self.visible = True
+        self.components = []
+        self.text = None
+        self.x = 0
+        self.y = 0
